+++ components/org.wso2.identity.webhook.common.event.handler/src/main/java/org/wso2/identity/webhook/common/event/handler/internal/util/EventConfigManager.java
@@ -43,8 +43,8 @@ import java.nio.file.Paths;
  */
 public class EventConfigManager {
 
-    private final ResourceConfig eventSchema;
     private static volatile EventConfigManager instance;
+    private final ResourceConfig eventSchema;
 
     EventConfigManager() throws IdentityEventServerException {
