+++ components/org.wso2.identity.webhook.wso2.event.handler/src/test/java/org/wso2/identity/webhook/wso2/event/handler/builder/WSO2LoginEventPayloadBuilderTest.java
@@ -106,7 +106,7 @@ public class WSO2LoginEventPayloadBuilderTest {
     @DataProvider(name = "successEventDataProvider")
     public Object[][] successEventDataProvider() {
 
-        return new Object[][] {
+        return new Object[][]{
                 {SAMPLE_USER_ID, SAMPLE_USERSTORE_NAME, SAMPLE_SP_ID, SAMPLE_SERVICE_PROVIDER, SAMPLE_TENANT_ID,
                         SAMPLE_TENANT_DOMAIN, SAMPLE_USER_REF, 1}
         };
@@ -145,7 +145,7 @@ public class WSO2LoginEventPayloadBuilderTest {
     @DataProvider(name = "failedEventDataProvider")
     public Object[][] failedEventDataProvider() {
 
-        return new Object[][] {
+        return new Object[][]{
                 {SAMPLE_USER_ID, SAMPLE_USERSTORE_NAME, SAMPLE_SP_ID, SAMPLE_SERVICE_PROVIDER, SAMPLE_TENANT_ID,
                         SAMPLE_TENANT_DOMAIN, SAMPLE_USER_REF, SAMPLE_ERROR_CODE, 2, SAMPLE_MESSAGE, SAMPLE_IDP, SAMPLE_AUTHENTICATOR}
         };
