+++ components/org.wso2.identity.webhook.wso2.event.handler/src/test/java/org/wso2/identity/webhook/wso2/event/handler/builder/util/TestUtils.java
@@ -44,6 +44,7 @@ public class TestUtils {
 
         builder = new ServiceURLBuilder() {
             String path = "/t/myorg";
+
             @Override
             public ServiceURLBuilder addPath(String... strings) {
